requirements.txt
bibtexparser==1.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
idna==3.7
iniconfig==2.0.0
joblib==1.4.2
nltk==3.8.1
packaging==24.0
pluggy==1.4.0
pycryptodome==3.20.0
pyparsing==3.1.2
pytest==8.1.1
python-slugify==8.0.4
regex==2024.4.28
requests==2.31.0
text-unidecode==1.3
tqdm==4.66.4
urllib3==2.2.1
